requirements.txt
# Core Backend Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
mysql-connector-python==8.2.0
PyMySQL==1.1.0

# NLP and ML
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Text Processing
spacy==3.7.2

# Voice Support
SpeechRecognition==3.10.1
pyttsx3==2.90
pyaudio==0.2.14

# Environment Variables
python-dotenv==1.0.0

# Security
bcrypt==4.1.2
